[gui2] Merge DesktopBrowserLauncher into Utilities
The utilities code appears to be a pass-through into the desktop code,
so just merge them all together and make the logic slightly less
complex.
