+++ code/src/java/pcgen/gui2/tools/Hyperactive.java
@@ -44,7 +44,7 @@ public final class Hyperactive implements HyperlinkListener
 
 			try
 			{
-				Utility.viewInBrowser(e.getURL());
+				DesktopBrowserLauncher.viewInBrowser(e.getURL());
 			}
 			catch (IOException t)
 			{
