+++ code/src/java/pcgen/resources/lang/LanguageBundle.properties
@@ -460,12 +460,14 @@ in_loadPartyInvalid=Invalid PCGen party file: {0}
 in_loadPartyNoRead=The party file {0} cannot be read.
 in_loadPartyNoSources=Could not find source files for party: {0}
 
+# Error Messages
+
+in_err_browser_err=Couldn't launch browser
+
 #About dialog
 in_abt_title=About PCGen
 
 in_abt_BD=Chief Code Monkey:
-
-in_abt_browser_err=Couldn't launch browser
 
 in_abt_code_mky=Code
